Separating out script filter stuff into a separate extention file.
